rename storj cli to storj-v2-cli
