+++ pkg/clientv2/env.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2018 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package client
+package clientv2
 
 import (
 	"strings"
